nginx-flask-mysql/backend/requirements.txt
Flask==1.1.1
mysql-connector==2.2.9
psycopg2
